See extended message
The program now writes the user's text after the first space to file. It
turns out cin doesn't work beyond the first space in a string, but
getline(cin, var) does. The problem with the program halting execution
after printing "Is that all, yes/no?" is fixed too. I just had to write
fflush(stdin) after the line which writes the user's input to the file.